+++ libraries/exoplayer_ima/src/main/java/androidx/media3/exoplayer/ima/AdTagLoader.java
@@ -29,6 +29,7 @@ import static java.lang.annotation.ElementType.TYPE_USE;
 import android.content.Context;
 import android.os.Handler;
 import android.os.SystemClock;
+import android.text.TextUtils;
 import android.view.ViewGroup;
 import androidx.annotation.IntDef;
 import androidx.annotation.Nullable;
@@ -73,6 +74,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -178,8 +180,8 @@ import java.util.Objects;
   /** Whether IMA has been notified that playback of content has finished. */
   private boolean sentContentComplete;
 
-  /** The MIME type of the ad pod that is next requested via an {@link AdEventType#LOADED} event. */
-  @Nullable private String pendingAdMimeType;
+  /** Mimetypes by AdInfo. Provided in the {@link AdEventType#LOADED} event */
+  private final Map<AdInfo, String> mimeTypeByAdInfo;
 
   // Fields tracking the player/loader state.
 
@@ -251,6 +253,7 @@ import java.util.Objects;
     }
     imaSdkSettings.setPlayerType(IMA_SDK_SETTINGS_PLAYER_TYPE);
     imaSdkSettings.setPlayerVersion(IMA_SDK_SETTINGS_PLAYER_VERSION);
+    mimeTypeByAdInfo = new HashMap<>();
     this.supportedMimeTypes = supportedMimeTypes;
     this.adTagDataSpec = adTagDataSpec;
     this.adsId = adsId;
@@ -789,7 +792,11 @@ import java.util.Objects;
         Log.i(TAG, message);
         break;
       case LOADED:
-        pendingAdMimeType = adEvent.getAd().getContentType();
+        String adMimeType = adEvent.getAd().getContentType();
+        if (!TextUtils.isEmpty(adMimeType)) {
+          // TODO: b/437870080 - Remove this once the SDK provides a mime type in the AdPodInfo.
+          setAdMimeType(adEvent.getAd().getAdPodInfo(), adMimeType);
+        }
         break;
       default:
         break;
@@ -997,10 +1004,11 @@ import java.util.Objects;
     }
 
     MediaItem.Builder adMediaItem = new MediaItem.Builder().setUri(adMediaInfo.getUrl());
-    if (pendingAdMimeType != null) {
-      adMediaItem.setMimeType(pendingAdMimeType);
-      pendingAdMimeType = null;
+    String adMimeType = mimeTypeByAdInfo.get(adInfo);
+    if (adMimeType != null) {
+      adMediaItem.setMimeType(adMimeType);
     }
+
     adPlaybackState =
         adPlaybackState.withAvailableAdMediaItem(
             adInfo.adGroupIndex, adInfo.adIndexInAdGroup, adMediaItem.build());
@@ -1268,6 +1276,13 @@ import java.util.Objects;
               AdLoadException.createForUnexpected(new RuntimeException(message, cause)),
               adTagDataSpec);
     }
+  }
+
+  private void setAdMimeType(AdPodInfo adPodInfo, String mimeType) {
+    int adGroupIndex = getAdGroupIndexForAdPod(adPodInfo);
+    int adIndexInAdGroup = adPodInfo.getAdPosition() - 1;
+    AdInfo adInfo = new AdInfo(adGroupIndex, adIndexInAdGroup);
+    mimeTypeByAdInfo.put(adInfo, mimeType);
   }
 
   private int getAdGroupIndexForAdPod(AdPodInfo adPodInfo) {